fix(lint): stop false-positive case-only category ID duplicates
checkDuplicateCategoryIds() lowercased IDs before comparing, so valid
distinct IDs like "Foo" and "foo" were flagged as hard-error duplicates.
The package writer (package/src/writer.js) dedupes categories with an
exact-case Map and SAFE_ID_REGEX permits mixed case, so those are two
legitimate, non-colliding IDs. Switch to exact-case duplicate detection
(error) and demote case-only collisions to a non-blocking warning.

Unlike categories, field IDs genuinely are looked up case-insensitively
elsewhere in the builder (payloadBuilder.ts's fieldNameToId map), so
checkDuplicateDetailIds() is intentionally left untouched.

test(ci): run lint parity tests in CI, gate deploy on Plugin checks

- Extract the GAS-boot vm harness (mock APIs, file collection, transpile)
  out of check-gas-boot.cjs into gasSimHarness.cjs so it can be reused.
- Add check-lint-parity.cjs, which boots the full source tree (including
  src/test/, unlike the boot probe) and runs runLintParityTests(). Wire it
  into the Plugin workflow as `test:lint-parity`. This was previously only
  runnable manually from the Apps Script editor, so it never ran in CI —
  exactly how the category ID bug above went uncaught.
- Fix two stale assertions in testLint.ts uncovered by actually running the
  suite: a severity check expecting "warning" where production now (and
  correctly) emits "error", and a duplicate-primaryLanguage message
  substring that didn't match current wording. Both were drift from the
  message/severity text changing without the test being updated.
- Gate deploy.yml on the Plugin workflow succeeding via workflow_run
  instead of triggering independently on the same push, so clasp push can
  no longer race ahead of failing or in-progress checks. Pin checkout to
  workflow_run's head_sha since it doesn't default to the triggering
  commit.
