+++ .github/workflows/deploy.yml
@@ -2,7 +2,12 @@ name: Deploy Script
 
 on:
   workflow_dispatch:
-  push:
+  # Gate on the Plugin checks workflow instead of triggering on the same push
+  # directly: a plain `push:` trigger here would race the Plugin workflow and
+  # could reach `clasp push` while checks are still running or have failed.
+  workflow_run:
+    workflows: ["Plugin"]
+    types: [completed]
     branches: [main, develop]
   release:
     types: [published]
@@ -13,6 +18,9 @@ concurrency:
 
 jobs:
   setup:
+    # For workflow_run, only proceed if the Plugin checks actually passed.
+    # workflow_dispatch and release runs have no upstream check run to gate on.
+    if: github.event_name != 'workflow_run' || github.event.workflow_run.conclusion == 'success'
     runs-on: ubuntu-latest
     outputs:
       matrix: ${{ steps.set-matrix.outputs.matrix }}
@@ -57,6 +65,11 @@ jobs:
     steps:
       - name: Checkout repo
         uses: actions/checkout@v4
+        with:
+          # workflow_run doesn't default to the commit that triggered the
+          # upstream Plugin run — pin explicitly so we deploy the exact
+          # commit that passed checks, not whatever HEAD happens to be.
+          ref: ${{ github.event.workflow_run.head_sha || github.sha }}
 
       - name: Install clasp
         run: sudo npm install @google/clasp@2.4.2 -g
