+++ .github/workflows/plugin.yml
@@ -48,3 +48,6 @@ jobs:
 
       - name: GAS boot probe
         run: bun run test:gas-boot
+
+      - name: Lint parity tests
+        run: bun run test:lint-parity
